+++ wasm/src/external_api/mod.rs
@@ -3,3 +3,4 @@ pub mod create_wallet;
 pub mod find_wallet;
 pub mod http;
 pub mod types;
+pub mod wallet_utils;